updated Stig scripts to allow download of artifacts from internet

+++ .common/artifacts/STIGs_Client/Apply-STIGs_Client.ps1
@@ -19,31 +19,6 @@ $Script:Name=[System.IO.Path]::GetFileNameWithoutExtension($Script:File)
 $virtualMachine = Get-WmiObject -Class Win32_ComputerSystem | Where-Object {$_.Model -match 'Virtual'}
 $osCaption = (Get-WmiObject -Class Win32_OperatingSystem).caption
 If ($osCaption -match 'Windows 11') { $osVersion = 11 } Else { $osVersion = 10 }
-$Script:Args = $null
-If ($ENV:PROCESSOR_ARCHITEW6432 -eq "AMD64") {
-    Try {
-
-        foreach($k in $MyInvocation.BoundParameters.keys)
-        {
-            switch($MyInvocation.BoundParameters[$k].GetType().Name)
-            {
-                "SwitchParameter" {if($MyInvocation.BoundParameters[$k].IsPresent) { $Script:Args += "-$k " } }
-                "String"          { $Script:Args += "-$k `"$($MyInvocation.BoundParameters[$k])`" " }
-                "Int32"           { $Script:Args += "-$k $($MyInvocation.BoundParameters[$k]) " }
-                "Boolean"         { $Script:Args += "-$k `$$($MyInvocation.BoundParameters[$k]) " }
-            }
-        }
-        If ($Script:Args) {
-            Start-Process -FilePath "$env:WINDIR\SysNative\WindowsPowershell\v1.0\PowerShell.exe" -ArgumentList "-File `"$($Script:FullName)`" $($Script:Args)" -Wait -NoNewWindow
-        } Else {
-            Start-Process -FilePath "$env:WINDIR\SysNative\WindowsPowershell\v1.0\PowerShell.exe" -ArgumentList "-File `"$($Script:FullName)`"" -Wait -NoNewWindow
-        }
-    }
-    Catch {
-        Throw "Failed to start 64-bit PowerShell"
-    }
-    Exit
-}
 
 [String]$Script:LogDir = "$($env:SystemRoot)\Logs\Configuration"
 If (-not(Test-Path -Path $Script:LogDir)) {
@@ -55,6 +30,130 @@ New-Item -Path $Script:TempDir -ItemType Directory -Force | Out-Null
 #endregion
 
 #region Functions
+
+Function Get-InternetFile {
+    [CmdletBinding()]
+    Param (
+        [Parameter(Mandatory = $true, Position = 0)]
+        [uri]$Url,
+        [Parameter(Mandatory = $true, Position = 1)]
+        [string]$OutputDirectory,
+        [Parameter(Mandatory = $false, Position = 2)]
+        [string]$OutputFileName
+    )
+
+    Begin {
+        $ProgressPreference = 'SilentlyContinue'
+        ## Get the name of this function and write header
+        [string]${CmdletName} = $PSCmdlet.MyInvocation.MyCommand.Name
+        Write-Verbose "Starting ${CmdletName} with the following parameters: $PSBoundParameters"
+    }
+    Process {
+
+        $start_time = Get-Date
+
+        If (!$OutputFileName) {
+            Write-Verbose "${CmdletName}: No OutputFileName specified. Trying to get file name from URL."
+            If ((split-path -path $Url -leaf).Contains('.')) {
+
+                $OutputFileName = split-path -path $url -leaf
+                Write-Verbose "${CmdletName}: Url contains file name - '$OutputFileName'."
+            }
+            Else {
+                Write-Verbose "${CmdletName}: Url does not contain file name. Trying 'Location' Response Header."
+                $request = [System.Net.WebRequest]::Create($url)
+                $request.AllowAutoRedirect=$false
+                $response=$request.GetResponse()
+                $Location = $response.GetResponseHeader("Location")
+                If ($Location) {
+                    $OutputFileName = [System.IO.Path]::GetFileName($Location)
+                    Write-Verbose "${CmdletName}: File Name from 'Location' Response Header is '$OutputFileName'."
+                }
+                Else {
+                    Write-Verbose "${CmdletName}: No 'Location' Response Header returned. Trying 'Content-Disposition' Response Header."
+                    $result = Invoke-WebRequest -Method GET -Uri $Url -UseBasicParsing
+                    $contentDisposition = $result.Headers.'Content-Disposition'
+                    If ($contentDisposition) {
+                        $OutputFileName = $contentDisposition.Split("=")[1].Replace("`"","")
+                        Write-Verbose "${CmdletName}: File Name from 'Content-Disposition' Response Header is '$OutputFileName'."
+                    }
+                }
+            }
+        }
+
+        If ($OutputFileName) { 
+            $wc = New-Object System.Net.WebClient
+            $OutputFile = Join-Path $OutputDirectory $OutputFileName
+            Write-Verbose "${CmdletName}: Downloading file at '$url' to '$OutputFile'."
+            Try {
+                $wc.DownloadFile($url, $OutputFile)
+                $time = (Get-Date).Subtract($start_time).Seconds
+                
+                Write-Verbose "${CmdletName}: Time taken: '$time' seconds."
+                if (Test-Path -Path $outputfile) {
+                    $totalSize = (Get-Item $outputfile).Length / 1MB
+                    Write-Verbose "${CmdletName}: Download was successful. Final file size: '$totalsize' mb"
+                    Return $OutputFile
+                }
+            }
+            Catch {
+                Write-Error "${CmdletName}: Error downloading file. Please check url."
+                Return $Null
+            }
+        }
+        Else {
+            Write-Error "${CmdletName}: No OutputFileName specified. Unable to download file."
+            Return $Null
+        }
+    }
+    End {
+        Write-Verbose "Ending ${CmdletName}"
+    }
+}
+
+Function Get-InternetUrl {
+    [CmdletBinding()]
+    Param (
+        [Parameter(Mandatory = $true, Position = 0)]
+        [uri]$Url,
+        [Parameter(Mandatory = $true, Position = 1)]
+        [string]$searchstring
+    )
+    Begin {
+        ## Get the name of this function and write header
+        [string]${CmdletName} = $PSCmdlet.MyInvocation.MyCommand.Name
+        Write-Verbose "${CmdletName}: Starting ${CmdletName} with the following parameters: $PSBoundParameters"
+    }
+    Process {
+
+        Try {
+            Write-Verbose -message "${CmdletName}: Now extracting download URL from '$Url'."
+            $HTML = Invoke-WebRequest -Uri $Url -UseBasicParsing
+            $Links = $HTML.Links
+            $ahref = $null
+            $ahref=@()
+            $ahref = ($Links | Where-Object {$_.href -like "*$searchstring*"}).href
+            If ($ahref.count -eq 0 -or $null -eq $ahref) {
+                $ahref = ($Links | Where-Object {$_.OuterHTML -like "*$searchstring*"}).href
+            }
+            If ($ahref.Count -eq 1) {
+                Write-Verbose -Message "${CmdletName}: Download URL = '$ahref'"
+                $ahref
+
+            }
+            Elseif ($ahref.Count -gt 1) {
+                Write-Verbose -Message "${CmdletName}: Download URL = '$($ahref[0])'"
+                $ahref[0]
+            }
+        }
+        Catch {
+            Write-Error "${CmdletName}: Error Downloading HTML and determining link for download."
+        }
+    }
+    End {
+        Write-Verbose -Message "${CmdletName}: Ending ${CmdletName}"
+    }
+}
 
 function New-Log {
     <#
@@ -161,18 +260,39 @@ New-Log -Path $Script:LogDir
 Write-Log -category Info -message "Starting '$PSCommandPath'."
 
 If (-not(Test-Path -Path "$env:SystemRoot\System32\Lgpo.exe")) {
-    $azipfiles = Get-ChildItem -Path $PSScriptRoot -filter '*.zip' -recurse
-    $lgpozip = $azipfiles[0].FullName
-    Write-Log -category Info -message "Expanding '$lgpozip' to '$DirTemp'."
-    expand-archive -path "$lgpozip" -DestinationPath $Script:TempDir -force
-    $algpoexe = Get-ChildItem -Path $Script:TempDir -filter 'lgpo.exe' -recurse
-    If ($algpoexe.count -gt 0) {
-        $lgpoexe = $algpoexe[0].FullName
-        Write-Log -category Info -message "Copying '$lgpoexe' to '$env:SystemRoot\system32'."
-        Copy-Item -Path $lgpoexe -Destination "$env:SystemRoot\System32" -force        
+    $LGPOZip = Join-Path -Path $PSScriptRoot -ChildPath 'LGPO.zip'
+    If (Test-Path $LPGOZip) {
+        Write-Log -category Info -message "Expanding '$LGPO' to '$Script:TempDir'."
+        expand-archive -path "$lgpozip" -DestinationPath $Script:TempDir -force
+        $algpoexe = Get-ChildItem -Path $Script:TempDir -filter 'lgpo.exe' -recurse
+        If ($algpoexe.count -gt 0) {
+            $lgpoexe = $algpoexe[0].FullName
+            Write-Log -category Info -message "Copying '$lgpoexe' to '$env:SystemRoot\system32'."
+            Copy-Item -Path $lgpoexe -Destination "$env:SystemRoot\System32" -force        
+        }
+    } Else {
+        $urlLGPO = 'https://download.microsoft.com/download/8/5/C/85C25433-A1B0-4FFA-9429-7E023E7DA8D8/LGPO.zip'
+        $fileLGPODownload = Get-InternetFile -Url $urlLGPO -OutputDirectory $Script:TempDir
+        $outputDir = Join-Path $Script:TempDir -ChildPath 'LGPO'
+        Expand-Archive -Path $fileLGPODownload -DestinationPath $outputDir
+        Remove-Item $fileLGPODownload -Force
+        $fileLGPO = (Get-ChildItem -Path $outputDir -file -Filter 'lgpo.exe' -Recurse)[0].FullName
+        Write-Output "Copying `"$fileLGPO`" to System32"
+        Copy-Item -Path $fileLGPO -Destination "$env:SystemRoot\System32" -Force
+        Remove-Item -Path $outputDir -Recurse -Force
     }
 }
-$stigZip = (Get-ChildItem -Path $PSScriptRoot -File -Filter '*STIG*.zip').FullName 
+
+$stigZip = Join-Path -Path $PSScriptRoot -ChildPath 'STIGs.zip'
+If (-not (Test-Path $stigZip)) {
+    #Download the STIG GPOs
+    $uriSTIGs = 'https://public.cyber.mil/stigs/gpo'
+    $uriGPODownload = Get-InternetUrl -Url $uriSTIGs -searchstring 'GPOs'
+    Write-Output "Downloading STIG GPOs from `"$uriGPODownload`"."
+    If ($uriGPODownload) {
+        $stigZip = Get-InternetFile -url $uriGPODownload -OutputDirectory $Script:TempDir
+    }
+} 
 
 Expand-Archive -Path $stigZip -DestinationPath $Script:TempDir -Force
 Write-Log -message "Copying ADMX and ADML files to local system."
